+++ GithubViewer/DVUserDetailTableViewController.m
@@ -50,7 +50,7 @@
         [self.tableView endUpdates];
         
     } failure:^(NSError *error, NSInteger statusCode) {
-        NSLog(@"%d: %@",  statusCode, [error localizedDescription]);
+        NSLog(@"%ld: %@",  (long)statusCode, [error localizedDescription]);
     }];
 }
